fix(security): external-review fixes — txid reuse, one-time login links, backup/proof/validation - v1.40.1
Six findings from an external review, all verified against the code and fixed:

- CRITICAL: BSC tx hashes are canonicalized to lowercase in the shared submit_reseller_payment
  before dedup + storage, so the same transfer as 0xABC… and 0xabc… can no longer bypass the
  unique index and settle multiple invoices.
- HIGH: portal login links are strictly one-time — each carries a random jti consumed atomically
  on exchange (new portal_login_nonce table; migration c8d2f1a4e6b9), so a leaked link can't be
  replayed for fresh sessions within its 15-minute TTL. The login page also strips ?t= immediately.
- HIGH: auto-backup without a passphrase now warns (owner message + log) that the unencrypted
  archive contains the system key. The key stays in the archive so cross-server restore keeps
  working (owner's explicit choice — warn, don't block).
- MEDIUM: the portal /pay/txid endpoint enforces the same chain-specific tx-hash format as the
  bot (shared validation in submit_reseller_payment), rejecting malformed/overlong values instead
  of 500ing on Postgres or storing junk review rows.
- MEDIUM: a portal screenshot proof that fails to save now notifies the owner text-only and asks
  the reseller to resend, instead of reporting success with a missing image (mirrors the bot).
- LOW: decimal excluded package sizes (1.5 GB) are matched exactly, not truncated to 1.

No new feature → PATCH. Full gate green: ruff, mypy (102 files), 235 backend tests, alembic
upgrade+check; frontend tsc+build; deploy-scripts. The portal_login_nonce migration auto-applies.

+++ backend/app/services/invoice_engine.py
@@ -115,7 +115,7 @@ def select_billable_roots(resellers: Sequence[Any]) -> list[Any]:
 
 
 # ----------------------------- billing -----------------------------
-def _excluded(usage_gb: float, excluded: set[int], free_threshold_gb: float) -> bool:
+def _excluded(usage_gb: float, excluded: set[float], free_threshold_gb: float) -> bool:
     """A config is a free test config when its quota is <= the free threshold
     (default 1 GB → 0.5 GB and 1 GB are free), OR it exactly matches an extra
     excluded size. Uses exact (not rounded) comparison so a real 1.3 GB package
@@ -132,7 +132,7 @@ def _excluded(usage_gb: float, excluded: set[int], free_threshold_gb: float) -> 
 def billable_gb_for_user(
     u: Any,
     period: Period,
-    excluded_usage_gb: set[int],
+    excluded_usage_gb: set[float],
     free_threshold_gb: float,
     panel_synced_at: dt.datetime | None = None,
     deleted_full_quota_over_gb: float = 0.0,
@@ -187,7 +187,7 @@ def compute_invoices(
     period: Period,
     *,
     default_price_per_gb: int,
-    excluded_usage_gb: set[int],
+    excluded_usage_gb: set[float],
     default_min_sale_toman: int = 0,
     free_threshold_gb: float = 1.0,
     panel_synced_at: dt.datetime | None = None,
